GEODE-4038: More size_t conversion. (#195)
- Refactors UserObjectSizer to use size_t.
- Standardizes on std::function for function pointers.
- Refactors DataIn/Output buffer methods to use size_t
- Refactors Properties::getSize
- Refactors Serializer::objectSize
- Removes unused SystemProperties::maxQueueSize
- Rename unused SystemProperties::javaConnectionPoolSize

+++ cppcache/include/geode/ResultSet.hpp
@@ -62,7 +62,7 @@ class _GEODE_EXPORT ResultSet : public SelectResults {
    * @returns A smart pointer to the item indexed.
    */
   virtual const std::shared_ptr<Serializable> operator[](
-      int32_t index) const override = 0;
+      size_t index) const override = 0;
 
   /**
    * Get a SelectResultsIterator with which to iterate over the items in the
