+++ cppcache/src/ResultSetImpl.cpp
@@ -35,7 +35,7 @@ int32_t ResultSetImpl::size() const {
 }
 
 const std::shared_ptr<Serializable> ResultSetImpl::operator[](
-    int32_t index) const {
+    size_t index) const {
   if (index >= m_resultSetVector->size()) {
     throw IllegalArgumentException("index out of bounds");
   }
